Implement write_read_repeated_start for I2C
This is needed for the accelerometer. The chip only allows reads from
registers following a repeated start following a write command. The
secret to this was section 27.8.7 of the atsam4l datasheet, which
revealed that to do a read following a write, you put the write command
into CMDR (as usual) and the read command into NCMDR, and then just
watch the TXRDY and RXRDY registers. We can decide later how to fold
this single method into the I2C interface