Implement magnetometer side of the acc/mag chip
Using the convention of 0 => both sensors (currently not implemented --
see below), 1 => accelerometer, 2 => magnetometer. Replying with both
sensors requires returning 6 values, which is currently beyond the
capabilities of the driver. We most likely do not want to change the
driver interface to account for this, so I need to figure out how to
share a buffer between userland and kernel using the Allow method.